Math Formula: Improve error handling + cleanup
When the code doesn't compile, the errors are now shown below the
formula. The positioning code for math nodes is now separated from the
add-on. The operator has been made modal so that the node width and
height can be calculated before positioning.

Checking for numbers now uses 'isdecimal()' rather than 'isdigit()',
because the latter allows unwanted characters.