Add fonts + basic formulas working
The next two big steps are macros and type checking. We need type
checking to ensure that no invalid args get passed to functions. It will
also allow operators like '*' to be smart (now it just assumes math
nodes). Getting basic macros working is also very nice because it means
that we can write macros for the vector_math and math functions.