Add multiline support in prompts.
Adding a new line can be done by pressing SHIFT + ENTER. For now this
means that the formula history has been removed because it was accessed
with the UP and DOWN arrows that are now used to navigate the multiple
lines.

This change also changes the prompt drawing to be more consistent
and less hacky. Concretely the formula is now actually drawn as
monospace, which makes it much easier to navigate, and draw the cursor
and errors.