Satisfy the type checker
Satisfy the 'basic' type checker by being more strict with type
annotations. This also uncovered some bugs. In the future stricter type
checking can be enabled, which will force adding of type annotations to
every function. Since this is already a huge change, that is left for
the future.

Fixes #9