Reuse created node groups in the same formula
When the node group created in a formula is called multiple times, it
will now be reused instead of being created from scratch.

Fixes #10

NOTE: If the node group was created in a previous formula, this doesn't
recycle the created node group. That is an improvement for a future
commit.