(MF): snapshot tests
We now do a form of "snapshot testing" for the resulting node tree. For
each of the test files, we store a simplified version of the node tree
to a file. When the tests are run again, the newly computed node tree
is compared against the old node tree. This way we can detect semantic
changes in the code, which might not have caused compiler errors.